requirements.txt
annotated-types==0.7.0
bottle==0.12.25
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coloredlogs==15.0.1
dimits==0.0.30a0
espeak_phonemizer==1.3.1
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.5.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
packaging==24.0
protobuf==5.27.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pyee==11.1.0
python-ffmpeg==2.0.12
PyYAML==6.0.1
requests==2.32.2
soundfile==0.12.1
sympy==1.12
tqdm==4.64.0
typing_extensions==4.12.0
urllib3==2.2.1
